chore: 타입스크립트 적용 중 	modified:   package-lock.json 	modified:   package.json 	modified:   src/components/Common/AddFolderButton.tsx 	modified:   src/components/Common/Button.tsx 	renamed:    src/components/Common/FlaotingAddFolderButton.tsx -> src/components/Common/FloatingAddFolderButton.tsx 	modified:   src/pages/FolderPage/FolderPage.tsx 	modified:   src/types.d.ts
