+++ src/types.d.ts
@@ -9,6 +9,6 @@ declare module '*.svg' {
 }
 
 declare module '*.png' {
-  const value: any;
-  export default value;
+  const src: string;
+  export default src;
 }
